added routing to basic page by user role

+++ src/main/java/com/queuesystem/request/Request.java
@@ -25,7 +25,7 @@ public class Request {
             strategy = GenerationType.SEQUENCE,
             generator = "request_sequence_sequence"
     )
-    private Long requestId;
+    private Integer requestId;
 
     @Column(nullable = false)
     private String filePath;
